DeckManagement, TagManagement views.
Cards belonging to decks and tags can now be managed separately from all
cards in existence. The management features are more restricted (only
unassign and disable) but sufficient when we take into account that
undecked card sort to the top of the card list in CardsOverview.

Also minor bugfix: going to the Start view always does a `loadAll`,
which should take care of the issue of editing decks and the changes not
appearing in DeckSelect.